ml-training/requirements.txt
jupyterlab>=4.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
holidays>=0.35
scipy>=1.11.0
boto3>=1.28.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
sentence-transformers>=2.2.2
transformers>=4.34.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
